+++ ydb/core/kafka_proxy/ya.make
@@ -58,6 +58,7 @@ PEERDIR(
     ydb/library/actors/core
     ydb/library/actors/protos
     ydb/core/base
+    ydb/core/persqueue/public/fetcher
     ydb/core/protos
     ydb/core/raw_socket
     ydb/services/persqueue_v1
