+++ ydb/core/persqueue/public/fetcher/fetch_request_actor.h
@@ -2,7 +2,7 @@
 
 #include <ydb/core/persqueue/public/constants.h>
 #include <ydb/core/persqueue/public/pq_rl_helpers.h>
-#include <ydb/library/actors/core/actor.h>
+#include <ydb/library/actors/core/actorsystem_fwd.h>
 #include <ydb/library/aclib/aclib.h>
 
 namespace NKikimr::NPQ {
