+++ ydb/core/persqueue/public/fetcher/fetch_request_ut.cpp
@@ -1,9 +1,10 @@
+#include "fetch_request_actor.h"
+
 #include <library/cpp/testing/unittest/registar.h>
-#include <ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils/ut_utils.h>
 #include <ydb/core/persqueue/events/internal.h>
-#include <ydb/core/persqueue/fetch_request_actor.h>
-#include <ydb/core/tx/scheme_board/cache.h>
 #include <ydb/core/testlib/tenant_runtime.h>
+#include <ydb/core/tx/scheme_board/cache.h>
+#include <ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils/ut_utils.h>
 
 namespace NKikimr::NPQ {
 using namespace NPersQueue;
