+++ ydb/core/persqueue/ut/ya.make
@@ -23,7 +23,6 @@ PEERDIR(
     ydb/core/persqueue/ut/common
     ydb/core/testlib/default
     ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils
-    ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils
 
     ydb/core/tx/schemeshard/ut_helpers
 )
@@ -47,7 +46,6 @@ SRCS(
     user_info_ut.cpp
     pqrb_describes_ut.cpp
     partition_scale_manager_graph_cmp_ut.cpp
-    fetch_request_ut.cpp
     utils_ut.cpp
     list_all_topics_ut.cpp
     cache_eviction_ut.cpp
